all tests have been consolidated into a single folder
