--- test/player.test.js
@@ -1,284 +1,0 @@
-/**
- * @jest-environment jsdom
- */
-
-// Integration tests
-
-const Player = require('../src/player')
-const Game = require('../src/game')
-const Render = require('../src/render')
-const fs = require('fs');
-
-describe('Player', () => {
-    beforeEach(() => {
-        render = new Render()
-        game = new Game(render)
-        document.body.innerHTML = fs.readFileSync('./index.html');
-    });
-
-    test('press left p1', () => {
-        const mockPressLeft = { key: "ArrowLeft"}
-        document.addEventListener = jest.fn((event, callback) => {
-            if (event === "keydown") {
-                callback(mockPressLeft)
-            }
-        })
-        const inputField = document.createElement('input');
-        document.body.appendChild(inputField);
-        inputField.focus();
-
-        
-        document.body.removeChild(inputField);
-        game.generateTetromino(0)
-        player = new Player(1, game);
-        
-        player.controls()
-
-        expect(mockPressLeft.key).toBe("ArrowLeft");
-        expect(game.grid).toEqual(
-            [
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 1, 1, 1, 1, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0]
-            ]
-        )
-    })
-
-    test('press left p1', () => {
-        const mockPressLeft = { key: "ArrowLeft"}
-        document.addEventListener = jest.fn((event, callback) => {
-            if (event === "keydown") {
-                callback(mockPressLeft)
-            }
-        })
-        const inputField = document.createElement('input');
-        document.body.appendChild(inputField);
-        inputField.focus();
-
-        
-        document.body.removeChild(inputField);
-        game.generateTetromino(1)
-        player = new Player(1, game);
-        
-        player.controls()
-        expect(mockPressLeft.key).toBe("ArrowLeft");
-        expect(game.grid).toEqual(
-            [
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 2, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 2, 2, 2, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0]
-            ]
-        )
-    })
-
-    test('press right p1', () => {
-        const mockPressRight = { key: "ArrowRight"}
-        document.addEventListener = jest.fn((event, callback) => {
-            if (event === "keydown") {
-                callback(mockPressRight)
-            }
-        })
-        const inputField = document.createElement('input');
-        document.body.appendChild(inputField);
-        inputField.focus();
-
-        
-        document.body.removeChild(inputField);
-        game.generateTetromino(2)
-        player = new Player(1, game);
-        
-        player.controls()
-        expect(mockPressRight.key).toBe("ArrowRight");
-        expect(game.grid).toEqual(
-            [
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 3, 0, 0, 0],
-                [0, 0, 0, 0, 3, 3, 3, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0]
-            ]
-        )
-    })
-
-    test('press left p2', () => {
-        const mockPressLeft = { key: "a"}
-        document.addEventListener = jest.fn((event, callback) => {
-            if (event === "keydown") {
-                callback(mockPressLeft)
-            }
-        })
-        const inputField = document.createElement('input');
-        document.body.appendChild(inputField);
-        inputField.focus();
-
-        
-        document.body.removeChild(inputField);
-        game.generateTetromino(3)
-        player = new Player(2, game);
-        
-        player.controls()
-        expect(mockPressLeft.key).toBe("a");
-        expect(game.grid).toEqual(
-            [
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 4, 4, 0, 0, 0, 0, 0],
-                [0, 0, 0, 4, 4, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0]
-            ]
-        )
-    })
-
-    test('press right p2', () => {
-        const mockPressRight = { key: "d"}
-        document.addEventListener = jest.fn((event, callback) => {
-            if (event === "keydown") {
-                callback(mockPressRight)
-            }
-        })
-        const inputField = document.createElement('input');
-        document.body.appendChild(inputField);
-        inputField.focus();
-
-        
-        document.body.removeChild(inputField);
-        game.generateTetromino(4)
-        player = new Player(2, game);
-        
-        player.controls()
-        expect(mockPressRight.key).toBe("d");
-        expect(game.grid).toEqual(
-            [
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 5, 5, 0, 0, 0],
-                [0, 0, 0, 0, 5, 5, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0]
-            ]
-        )
-    })
-
-    test('press left p1 twice', () => {
-        const mockPressLeft = { key: "ArrowLeft"}
-        document.addEventListener = jest.fn((event, callback) => {
-            if (event === "keydown") {
-                callback(mockPressLeft)
-            }
-        })
-        const inputField = document.createElement('input');
-        document.body.appendChild(inputField);
-        inputField.focus();
-
-        
-        document.body.removeChild(inputField);
-        game.generateTetromino(0)
-        player = new Player(1, game);
-        
-        player.controls()
-        player.controls()
-        expect(mockPressLeft.key).toBe("ArrowLeft");
-        expect(game.grid).toEqual(
-            [
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 1, 1, 1, 1, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                [0, 0, 0, 0, 0, 0, 0, 0, 0, 0]
-            ]
-        )
-    })
-})
